Refactor HTML structure and add CSS styles (#1)
* Refactor HTML structure and add CSS styles

This commit refactors the HTML structure of the single_page.html file by adding a head section with a title and a body section with a container div. It also adds CSS styles to the style.css file to improve the visual appearance of the web page.

* Delete .vscode/settings.json

* Update single_page.html

* Delete style.css

* Update single_page.html

Update to have complete and easy access on phone and tablet.

* Update single_page.html

Update to align connect button and color picker horizontally.